requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
openai
python-dotenv
